feat(sl): change API so that step no. can increase
